+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Support LSP 3.16 file operations: `workspace/willRenameFiles`, `workspace/didRenameFiles`, 
+`workspace/willCreateFiles`, `workspace/didCreateFiles`, `workspace/willDeleteFiles`, `workspace/didDeleteFiles`.
+
 ## v0.0.1
 
 - First release